core/mvcc: Move source code to module
