+++ core/Cargo.toml
@@ -70,6 +70,8 @@ limbo_time = { path = "../extensions/time", optional = true, features = ["static
 miette = "7.4.0"
 strum = "0.26"
 parking_lot = "0.12.3"
+tracing = "0.1.41"
+crossbeam-skiplist = "0.1.3"
 
 [build-dependencies]
 chrono = "0.4.38"
